Use Int as Player id

+++ server/src/main/kotlin/dev/chargedbyte/reaktor_summer_2022/feature/player/PlayerDto.kt
@@ -1,3 +1,3 @@
 package dev.chargedbyte.reaktor_summer_2022.feature.player
 
-data class PlayerDto(val id: Long?, val name: String)
+data class PlayerDto(val id: Int?, val name: String)
